+++ src/Components/LeaveComments/index.tsx
@@ -1,10 +1,8 @@
-import React, { useMemo, useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
-import * as I from "../../Asset/SVG/index";
+import { useEffect, useMemo, useState } from "react";
 import * as S from "./style";
 import IssueBoxPresenter from "../IssueBox";
-import Config from "Constants/Config.json";
 import Table from "Api/table";
+import LeftBox from "Components/LeftBox";
 
 interface list {
   uncomfortableIdx: number;
@@ -26,19 +24,6 @@ const useGetTable = () => {
   return list;
 };
 
-const useWriteTable = () => {
-  const writeTable = async (
-    content: string,
-    setContent: React.Dispatch<React.SetStateAction<string>>,
-  ) => {
-    setContent("");
-    window.location.reload();
-    return await Table.writeTable(content);
-  };
-
-  return writeTable;
-};
-
 const useShuffle = () => {
   const shuffle = (list: Array<list>) => {
     return list.sort(() => Math.random() - 0.5);
@@ -49,11 +34,7 @@ const useShuffle = () => {
 
 const LeaveCommentsPage: React.FC = () => {
   const list = useGetTable();
-  const history = useHistory();
   const shuffle = useShuffle();
-
-  const [content, setContent] = useState("");
-  const [length, setLength] = useState(0);
 
   const provList = useMemo(() => shuffle(list), [list]);
 
@@ -62,47 +43,12 @@ const LeaveCommentsPage: React.FC = () => {
   const list2 = provList.slice(list.length * (2 / 4), list.length * (3 / 4));
   const list1 = provList.slice(list.length * (3 / 4), list.length);
 
-  const tryWriteTable = useWriteTable();
-
-  const textAreaOnchange = (e: any) => {
-    setLength(e.target.value.length);
-    setContent(e.target.value);
-  };
+  const heading: string[] = ["학교가 불편할 때는,", "언제였나요?"];
+  const explanation: string[] = [];
 
   return (
     <S.LeaveCommentsBox>
-      <S.LeftBox>
-        <S.H1>
-          학교가 불편할 때는,
-          <br />
-          언제였나요?
-        </S.H1>
-        <S.Form>
-          <textarea
-            placeholder="자유롭게 의견을 남겨주세요."
-            onChange={textAreaOnchange}
-            value={content}
-          />
-          <S.Btn
-            Length={length}
-            onClick={() => {
-              length >= 8 && tryWriteTable(content, setContent);
-            }}
-          >
-            등 록
-          </S.Btn>
-        </S.Form>
-        <S.RankBtn onClick={() => history.push(Config.LINK.RANK)}>
-          <span>
-            많은 학생들이 공감한
-            <br /> 불편함은 무엇일까요?
-          </span>
-          <S.LinkWrapper>
-            <span>Top 30 보러가기</span>
-            <I.RightArrow />
-          </S.LinkWrapper>
-        </S.RankBtn>
-      </S.LeftBox>
+      <LeftBox heading={heading} explanation={explanation}/>
       <S.RightBox>
         <S.IssueBoxWrapper>
           {list1.map((table: list, index) => (
